update google chrome download URL
Starting on 7/23/24, our Heroku builds were failing, because it could not retrieve the specified chrome version
using the dl.google.com URL structure. While I couldn't find a specific changelog that pointed to this URL no
longer working, Andy from RM helped me find the list of supported Chrome versions and their download paths here:
https://github.com/GoogleChromeLabs/chrome-for-testing/blob/main/data/known-good-versions-with-downloads.json#L4977

Since we were using a forked copy of the heroku-buildpack-google-chrome repo that pointed to the old URL structure,
I created a forked copy under the Maverick-Solutions-NC org and updated the path to point to the right one.
However, it does point to a .zip file vs. a .deb file, so I copied some commands from the heroku version of the
buildpack to unzip the file: https://github.com/heroku/heroku-buildpack-chromedriver/blob/master/bin/compile#L66C1-L73C1

Note that we will probably want to host our own version of the v116, so that in the future, if this version ever
gets deprecated, we can still have access to it through our own hosted version. However, a better short-term goal
is to figure out how to update our Chrome version to a more recent one, when time permits.